Unify analysis reporting kernel
