--- .gitkeep
@@ -1,1 +1,0 @@
-# .gitkeep file auto-generated at 2026-04-20T18:37:53.802Z for PR creation at branch issue-85-10187f742854 for issue https://github.com/netkeep80/repo-guard/issues/85
