+++ tests/test-pipeline.mjs
@@ -1,8 +1,14 @@
 import { strict as assert } from "node:assert";
+import { mkdtempSync, rmSync, writeFileSync } from "node:fs";
+import { tmpdir } from "node:os";
+import { join, resolve } from "node:path";
 import { buildPolicyFacts } from "../src/facts/input.mjs";
+import { createIntegrationAnalysisReport } from "../src/integration-validator.mjs";
 import { runPolicyPipeline } from "../src/runtime/pipeline.mjs";
 
 let failures = 0;
+const __dirname = new URL(".", import.meta.url).pathname;
+const projectRoot = resolve(__dirname, "..");
 
 function expect(label, actual, expected) {
   try {
@@ -83,6 +89,56 @@ function buildEquivalentFacts(extra = {}) {
   });
 }
 
+function expectCanonicalEnvelope(label, report, command) {
+  const canonicalKeys = [
+    "command",
+    "mode",
+    "ok",
+    "result",
+    "passed",
+    "failed",
+    "violations",
+    "advisoryWarnings",
+    "warnings",
+    "violationCount",
+    "exitCode",
+    "ruleResults",
+    "hints",
+    "repositoryRoot",
+  ];
+  expect(`${label} command`, report.command, command);
+  expect(`${label} canonical keys`, canonicalKeys.every((key) =>
+    Object.prototype.hasOwnProperty.call(report, key)
+  ), true);
+  expect(`${label} ruleResults array`, Array.isArray(report.ruleResults), true);
+  expect(`${label} violations array`, Array.isArray(report.violations), true);
+  expect(`${label} hints array`, Array.isArray(report.hints), true);
+}
+
+function makeIntegrationKernelRepo() {
+  const dir = mkdtempSync(join(tmpdir(), "repo-guard-kernel-"));
+  const policy = {
+    policy_format_version: "0.3.0",
+    repository_kind: "tooling",
+    paths: {
+      forbidden: [],
+      canonical_docs: ["README.md"],
+      governance_paths: ["repo-policy.json"],
+    },
+    diff_rules: {
+      max_new_docs: 5,
+      max_new_files: 5,
+      max_net_added_lines: 500,
+    },
+    content_rules: [],
+    cochange_rules: [],
+  };
+
+  writeFileSync(join(dir, "repo-policy.json"), JSON.stringify(policy, null, 2));
+  writeFileSync(join(dir, "README.md"), "# Test\n");
+  return dir;
+}
+
 console.log("\n--- shared policy pipeline normalizes facts and checks ---");
 {
   const facts = buildEquivalentFacts();
@@ -120,9 +176,18 @@ console.log("\n--- equivalent command inputs share one result shape ---");
     contractSource: "pr body",
     initialChecks: [{ name: "change-contract", check: { ok: true } }],
   });
+  const integrationRepo = makeIntegrationKernelRepo();
+  const validateIntegrationStyle = createIntegrationAnalysisReport({
+    packageRoot: projectRoot,
+    repoRoot: integrationRepo,
+    enforcementMode: null,
+  }, { format: "json" });
   const checkDiffFacts = buildEquivalentFacts();
   const checkPrFacts = buildEquivalentFacts({ mode: "check-pr", contractSource: "pr body" });
 
+  expectCanonicalEnvelope("check-diff report", checkDiffStyle, "check-diff");
+  expectCanonicalEnvelope("check-pr report", checkPrStyle, "check-pr");
+  expectCanonicalEnvelope("validate-integration report", validateIntegrationStyle, "validate-integration");
   expect("equivalent facts keep mode-specific provenance", {
     mode: checkPrFacts.mode,
     contractSource: checkPrFacts.contractSource,
@@ -136,6 +201,7 @@ console.log("\n--- equivalent command inputs share one result shape ---");
     checkPrStyle.violations.map((violation) => violation.rule),
     checkDiffStyle.violations.map((violation) => violation.rule)
   );
+  rmSync(integrationRepo, { recursive: true });
 }
 
 console.log("\n--- check-pr style pipeline evaluates size rules ---");
